+++ src/utilities.h
@@ -78,7 +78,7 @@ static version_t get_version(void)
 static void print_logo(void)
 {
 	const version_t version = get_version();
-	fprintf(stderr, "\nMHash384 v%u.%u.%u, simple fast portable header-only hashing library [%s]\n", version.major, version.minor, version.patch, __DATE__);
+	fprintf(stderr, "\nMHash384 v%u.%u.%u, simple fast portable header-only hashing library [%s]\n", (unsigned int)version.major, (unsigned int)version.minor, (unsigned int)version.patch, __DATE__);
 	fprintf(stderr, "Copyright(c) 2016-2017 LoRd_MuldeR <mulder2@gmx.de>, released under the MIT License.\n\n");
 }
 
@@ -148,7 +148,7 @@ static int parse_arguments(param_t *param, int argc, CHAR *argv[])
 			else if (!STRICMP(argv[i], T("-v")) || !STRICMP(argv[i], T("--version")))
 			{
 				const version_t version = get_version();
-				printf("mhash-384 version %u.%u.%u (built %s)\n", version.major, version.minor, version.patch, __DATE__);
+				printf("mhash-384 version %u.%u.%u (built %s)\n", (unsigned int)version.major, (unsigned int)version.minor, (unsigned int)version.patch, __DATE__);
 				return 0;
 			}
 			else if ((argv[i][0] == T('-')) && (argv[i][1] != T('\0')))
